+++ src/rendering/mod.rs
@@ -1,2 +1,3 @@
 pub mod display;
 pub mod scroll_selector;
+pub mod search;